refactor: add constants to api and styled module to all components
