+++ src/pages/Home/Home.tsx
@@ -1,16 +1,23 @@
 import { JSX, useState } from 'react';
 import './Home.css';
+import {
+  CardListWrapper,
+  CardImageSmall,
+} from '../../components/SmallCardList/styled';
+import { Loader } from '../../components/CardList/styled';
+import {
+  ArtInfo,
+  URL_ARTWORK,
+  URL_IMAGE,
+  URL_SEARCH,
+} from '../../constants/api';
 import Header from '../../components/Header/Header';
 import SearchBar from '../../components/SearchBar/SearchBar';
 import GallerySection from '../../components/GallerySection/GallerySection';
 import CardList from '../../components/CardList/CardList';
-import SmallCardList, {
-  CardListWrapper,
-  CardImageSmall,
-} from '../../components/SmallCardList/SmallCardList';
+import SmallCardList from '../../components/SmallCardList/SmallCardList';
 import SmallCard from '../../components/SmallCard/SmallCard';
 import Footer from '../../components/Footer/Footer';
-import { ArtInfo, Loader } from '../../components/CardList/CardList';
 
 function Home(): JSX.Element {
   const [searchResults, setSearchResults] = useState<ArtInfo[]>([]);
@@ -20,16 +27,14 @@ function Home(): JSX.Element {
     setLoading(true);
 
     try {
-      const response = await fetch(
-        `https://api.artic.edu/api/v1/artworks/search?q=${searchTerm}&limit=9`,
-      );
+      const response = await fetch(URL_SEARCH({ searchTerm }));
       const result = await response.json();
       const artworks = result.data;
 
       const detailedArtworks = await Promise.all(
         artworks.map(async (artwork: { id: number }) => {
           const detailResponse = await fetch(
-            `https://api.artic.edu/api/v1/artworks/${artwork.id}`,
+            URL_ARTWORK({ artworkId: artwork.id }),
           );
           const detailResult = await detailResponse.json();
           return detailResult.data;
@@ -71,7 +76,7 @@ function Home(): JSX.Element {
                       is_public_domain={item.is_public_domain}
                     >
                       <CardImageSmall
-                        image_url={`https://www.artic.edu/iiif/2/${item.image_id}/full/843,/0/default.jpg`}
+                        image_url={URL_IMAGE({ imageId: item.image_id })}
                       />
                     </SmallCard>
                   ))}
